Clean up Central's scannedPeripherals dict.
When COVIDSafe removed the timer code, it also removed the reset of this
dictionary, so currently it grows and grows in size. It probably doesn't get
large enough to cause real problems (e.g. the OS killing the app), but for
completeness also reset it on a memory warning. By my scrap paper calculations,
a group of 50 people (with identifiers changing every 15m) will get it to ~200kb
in 2 hours, so potentially over days it could cause issues currently.